added api routes

+++ rangers_shop/forms.py
@@ -22,6 +22,8 @@ class ProductForm(FlaskForm):
     name = StringField('Exercise Name', validators=[ DataRequired() ] )
     image = StringField('Img url **Optional')
     muscle = StringField('Target Muscle **Optional')
-    sets = IntegerField('Sets', validators=[ DataRequired() ])
+    sets = DecimalField('Sets', validators=[ DataRequired() ])
     reps = IntegerField('Reps', validators=[ DataRequired() ])
     submit = SubmitField('Submit')
+
+
